Move LCOW & WCOW layer management functions to their own file
As we refactor the layer management code, it is easier to keep LCOW & WCOW layer management code
in their own separate files.

(cherry picked from commit 47a65a15819fb13236d780ec057f77eaddc8f367)
